chore: drop shanbay & rename shanbayword to shanbay
